feat: provide a way to skip the focusser
There was no easy way to configure if the focusser should be skipped
after selecting an item. The fact that the focusser is always focused
after selecting an item is especially problematic on mobile devices,
since this leads to the keyboard being shown, which might not be
desirable.

This commit provides the skip-focusser option to configure this
behavior on the ui-select directive.

Fixes #869, #401, #818, #603, #432 (partially)